move the function to xmlutils to keep the consisntency
